pyproject.toml
[project]
name = "jojo-code"
version = "0.1.0"
description = "A coding agent powered by jojo AI - TypeScript CLI + Python LangGraph Core"
readme = "README.md"
license = "MIT"
authors = [
    { name = "afine907", email = "1049905106@qq.com" }
]
requires-python = ">=3.11"
keywords = [
    "agent",
    "langgraph",
    "langchain",
    "coding-assistant",
    "llm",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langchain-anthropic>=0.2",
    "pydantic>=2",
    "pydantic-settings>=2",
    "tiktoken>=0.5",
    "python-dotenv>=1",
    "duckduckgo-search>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4",
    "mypy>=1",
    "ruff>=0.4",
    "python-dotenv>=1",
    "pexpect>=4.9",  # E2E CLI testing
]

[project.urls]
Homepage = "https://github.com/afine907/jojo-code"
Documentation = "https://github.com/afine907/jojo-code#readme"
Repository = "https://github.com/afine907/jojo-code"
Issues = "https://github.com/afine907/jojo-code/issues"

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4"]
ignore = ["E402"]  # Module level import not at top of file (needed for load_dotenv)

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "e2e: 真实 API 调用的端到端测试",
    "longcat: 使用 LongCat API 的测试",
    "slow: 运行较慢的测试",
]
